#2521: Merged trunk changes into branch.
git-svn-id: https://src.heuristiclab.com/svn/core/branches/2521_ProblemRefactoring@18086 2abd9481-f8db-48e9-bd25-06bc13291c1b

+++ HeuristicLab.Visualization.ChartControlsExtensions/3.3/ChartUtil.cs
@@ -87,8 +87,8 @@ namespace HeuristicLab.Visualization.ChartControlsExtensions {
 
     private static double RoundDown(this double value, int decimalRank) {
       if (decimalRank > 0) {
-        var floor = (int)Math.Floor(value);
-        var pow = (int)Math.Pow(10, decimalRank);
+        var floor = Math.Floor(value);
+        var pow = Math.Pow(10, decimalRank);
         var mod = Mod(floor, pow);
         return floor - mod;
       }
@@ -97,8 +97,8 @@ namespace HeuristicLab.Visualization.ChartControlsExtensions {
 
     private static double RoundUp(this double value, int decimalRank) {
       if (decimalRank > 0) {
-        var ceil = (int)Math.Ceiling(value);
-        var pow = (int)Math.Pow(10, decimalRank);
+        var ceil = Math.Ceiling(value);
+        var pow = Math.Pow(10, decimalRank);
         var mod = Mod(ceil, pow);
         return ceil - mod + pow;
       }
